pyproject.toml
[tool.poetry]
name = "hi-hello-fastapi"
version = "0.1.0"
description = "hi-hello-fastapi"
authors = ["kookong2 <hy010309@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
sqlalchemy = "^2.0.36"
pymysql = "^1.1.1"
python-dotenv = "^1.0.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
